[IMP] estate:  Added model and view inheritance, Link Module, Kanban view
Added model inheritance in the inherited_res_users.py file, and view inheritance
in the inherited_res_users_views.xml file. Added link module estate_account and
added 'estate' and 'account' modules in its dependency. Inherited
estate.property model in the estate_property.py file in estate_account module.
Created invoice and added invoice lines. Added kanban view in the
estate_property_views.xml file and added default grouping and prevented drag and
drop.

+++ estate/models/estate_property_offer.py
@@ -32,7 +32,7 @@ class EstatePropertyOffer(models.Model):
     @api.model
     def create(self, vals):
         self.env['estate.property'].browse(vals['property_id']).state = 'offer received'
-        if float(self.env['estate.property.offer'].browse(vals['price'])) < self.env['estate.property'].browse(vals['property_id']).best_offer:
+        if int(self.env['estate.property.offer'].browse(vals['price'])) < self.env['estate.property'].browse(vals['property_id']).best_offer:
             raise exceptions.UserError(f"The offer must be higher than {self.env['estate.property'].browse(vals['property_id']).best_offer}")
         return super(EstatePropertyOffer, self).create(vals)
